Options pour la fin du parcours aussi
